Split case subject lowering modules
- Move Bool/Int/Float/String case subject lowering under `case/subject`.
- Keep `case.rs` as the case facade for subject validation and shared error constructors.
- Rehome subject-owned tests beside the new subject module without changing case behavior.
