+++ src/planner/expression/case/subject/int.rs
@@ -1,7 +1,6 @@
-use super::{
-    case_return_type, invalid_case_shape, single_case_pattern, unsupported_case,
-    validate_clause_shape,
-};
+use super::super::super::plan_int_expr;
+use super::super::{invalid_case_shape, unsupported_case};
+use super::{case_return_type, single_case_pattern, validate_clause_shape};
 use crate::plan::{BoolExpr, Expr, ExprKind, IntCaseBranches, IntExpr, ValueType};
 use crate::planner::context::PlanContext;
 use crate::planner::error::{InvalidCaseShapeReason, PlanError, UnsupportedCaseReason};
@@ -17,7 +16,7 @@ pub(super) fn plan(
     clauses: Vec<TypedClause>,
     context: &mut PlanContext<'_>,
 ) -> Result<Expr, PlanError> {
-    let subject = super::super::plan_int_expr(subject, context)?;
+    let subject = plan_int_expr(subject, context)?;
     let return_type = case_return_type(type_.as_ref())?;
     for clause in &clauses {
         validate_clause_shape(clause)?;
@@ -1132,7 +1131,7 @@ pub fn main() {
     #[test]
     fn reject_margin_int_case_pattern_shapes() {
         let mut variable_type_mismatch = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut variable_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Variable {
@@ -1151,7 +1150,7 @@ pub fn main() {
         );
 
         let mut discard_type_mismatch = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut discard_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[1].pattern[0] = Pattern::Discard {
@@ -1169,7 +1168,7 @@ pub fn main() {
         );
 
         let mut invalid_pattern = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut invalid_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Invalid {
@@ -1186,7 +1185,7 @@ pub fn main() {
         );
 
         let mut pattern_type_mismatch = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut pattern_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::String {
@@ -1203,7 +1202,7 @@ pub fn main() {
         );
 
         let mut assign_invalid_pattern = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut assign_invalid_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Assign {
@@ -1224,7 +1223,7 @@ pub fn main() {
         );
 
         let mut assign_type_mismatch = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut assign_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Assign {
@@ -1245,7 +1244,7 @@ pub fn main() {
         );
 
         let mut empty_pattern = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut empty_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern.clear();
@@ -1259,7 +1258,7 @@ pub fn main() {
         );
 
         let mut case_type_mismatch = compile_int_case_module();
-        let (case_type, _, _) = super::super::expect_case_statement_mut(
+        let (case_type, _, _) = super::super::super::expect_case_statement_mut(
             &mut case_type_mismatch.definitions.functions[0].body[0],
         );
         *case_type = type_::bool();
@@ -1273,7 +1272,7 @@ pub fn main() {
         );
 
         let mut missing_fallback_pattern = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut missing_fallback_pattern.definitions.functions[0].body[0],
         );
         clauses.pop();
@@ -1313,7 +1312,8 @@ fn add_one(value: Int) {
             })
             .map(|function| &mut function.body)
             .expect("expected main function");
-        let (_, _, clauses) = super::super::expect_assignment_case_statement_mut(&mut body[0]);
+        let (_, _, clauses) =
+            super::super::super::expect_assignment_case_statement_mut(&mut body[0]);
         clauses.pop();
         assert_eq!(
             plan_module(missing_function_fallback_pattern),
@@ -1328,8 +1328,9 @@ fn add_one(value: Int) {
     #[test]
     fn reject_margin_int_case_guard_must_be_bool() {
         let mut module = compile_int_case_module();
-        let (_, _, clauses) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
             location: dummy_span(),
             value: "1".into(),
@@ -1350,7 +1351,7 @@ fn add_one(value: Int) {
     #[test]
     fn reject_margin_guarded_int_case_pattern_shapes() {
         let mut empty_pattern = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut empty_pattern.definitions.functions[0].body[0],
         );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
@@ -1369,7 +1370,7 @@ fn add_one(value: Int) {
         );
 
         let mut pattern_type_mismatch = compile_int_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut pattern_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
@@ -1394,8 +1395,9 @@ fn add_one(value: Int) {
     #[test]
     fn reject_margin_int_case_subject_type_mismatch() {
         let mut module = compile_int_case_module();
-        let (_, subjects, _) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (_, subjects, _) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         subjects[0] = gleam_core::ast::TypedExpr::String {
             location: dummy_span(),
             type_: type_::int(),
@@ -1416,8 +1418,9 @@ fn add_one(value: Int) {
     #[test]
     fn reject_margin_int_case_expr_type_mismatch() {
         let mut module = compile_int_case_module();
-        let (type_, _, _) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (type_, _, _) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         *type_ = type_::bit_array();
         assert_eq!(plan_module(module), Err(case_branch_return_type_mismatch()));
 
@@ -1633,8 +1636,9 @@ fn stringify(value: Int) {
             })
             .map(|function| &mut function.body)
             .expect("expected main function");
-        let replacement = super::super::expect_expression_statement(&body[1]).clone();
-        let (_, _, clauses) = super::super::expect_assignment_case_statement_mut(&mut body[0]);
+        let replacement = super::super::super::expect_expression_statement(&body[1]).clone();
+        let (_, _, clauses) =
+            super::super::super::expect_assignment_case_statement_mut(&mut body[0]);
         clauses[1].then = replacement;
 
         assert_eq!(
