+++ src/planner/expression/case/subject/string.rs
@@ -1,7 +1,6 @@
-use super::{
-    case_return_type, invalid_case_shape, single_case_pattern, unsupported_case,
-    validate_clause_shape,
-};
+use super::super::super::plan_string_expr;
+use super::super::{invalid_case_shape, unsupported_case};
+use super::{case_return_type, single_case_pattern, validate_clause_shape};
 use crate::plan::{BoolExpr, Expr, ExprKind, StringCaseBranches, StringExpr, ValueType};
 use crate::planner::context::PlanContext;
 use crate::planner::error::{InvalidCaseShapeReason, PlanError, UnsupportedCaseReason};
@@ -16,7 +15,7 @@ pub(super) fn plan(
     clauses: Vec<TypedClause>,
     context: &mut PlanContext<'_>,
 ) -> Result<Expr, PlanError> {
-    let subject = super::super::plan_string_expr(subject, context)?;
+    let subject = plan_string_expr(subject, context)?;
     let return_type = case_return_type(type_.as_ref())?;
     for clause in &clauses {
         validate_clause_shape(clause)?;
@@ -892,7 +891,7 @@ pub fn main() {
     #[test]
     fn reject_margin_string_case_pattern_shapes() {
         let mut alternative_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut alternative_pattern.definitions.functions[0].body[0],
         );
         clauses[0].alternative_patterns.push(vec![Pattern::String {
@@ -907,7 +906,7 @@ pub fn main() {
         );
 
         let mut variable_type_mismatch = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut variable_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Variable {
@@ -926,7 +925,7 @@ pub fn main() {
         );
 
         let mut discard_type_mismatch = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut discard_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[1].pattern[0] = Pattern::Discard {
@@ -944,7 +943,7 @@ pub fn main() {
         );
 
         let mut invalid_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut invalid_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Invalid {
@@ -961,7 +960,7 @@ pub fn main() {
         );
 
         let mut pattern_type_mismatch = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut pattern_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Int {
@@ -979,7 +978,7 @@ pub fn main() {
         );
 
         let mut assign_invalid_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut assign_invalid_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Assign {
@@ -1000,7 +999,7 @@ pub fn main() {
         );
 
         let mut assign_type_mismatch = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut assign_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].pattern[0] = Pattern::Assign {
@@ -1022,7 +1021,7 @@ pub fn main() {
         );
 
         let mut empty_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut empty_pattern.definitions.functions[0].body[0],
         );
         clauses[0].pattern.clear();
@@ -1036,7 +1035,7 @@ pub fn main() {
         );
 
         let mut case_type_mismatch = compile_string_case_module();
-        let (case_type, _, _) = super::super::expect_case_statement_mut(
+        let (case_type, _, _) = super::super::super::expect_case_statement_mut(
             &mut case_type_mismatch.definitions.functions[0].body[0],
         );
         *case_type = type_::bool();
@@ -1050,7 +1049,7 @@ pub fn main() {
         );
 
         let mut missing_fallback_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut missing_fallback_pattern.definitions.functions[0].body[0],
         );
         clauses.pop();
@@ -1090,7 +1089,8 @@ fn return_value(value: String) {
             })
             .map(|function| &mut function.body)
             .expect("expected main function");
-        let (_, _, clauses) = super::super::expect_assignment_case_statement_mut(&mut body[0]);
+        let (_, _, clauses) =
+            super::super::super::expect_assignment_case_statement_mut(&mut body[0]);
         clauses.pop();
         assert_eq!(
             plan_module(missing_function_fallback_pattern),
@@ -1105,7 +1105,7 @@ fn return_value(value: String) {
     #[test]
     fn reject_margin_guarded_string_case_pattern_shapes() {
         let mut empty_pattern = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut empty_pattern.definitions.functions[0].body[0],
         );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
@@ -1124,7 +1124,7 @@ fn return_value(value: String) {
         );
 
         let mut pattern_type_mismatch = compile_string_case_module();
-        let (_, _, clauses) = super::super::expect_case_statement_mut(
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
             &mut pattern_type_mismatch.definitions.functions[0].body[0],
         );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
@@ -1150,8 +1150,9 @@ fn return_value(value: String) {
     #[test]
     fn reject_margin_string_case_guard_must_be_bool() {
         let mut module = compile_string_case_module();
-        let (_, _, clauses) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (_, _, clauses) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         clauses[0].guard = Some(ClauseGuard::Constant(Constant::Int {
             location: dummy_span(),
             value: "1".into(),
@@ -1172,8 +1173,9 @@ fn return_value(value: String) {
     #[test]
     fn reject_margin_string_case_subject_type_mismatch() {
         let mut module = compile_string_case_module();
-        let (_, subjects, _) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (_, subjects, _) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         subjects[0] = gleam_core::ast::TypedExpr::Int {
             location: dummy_span(),
             type_: type_::string(),
@@ -1195,8 +1197,9 @@ fn return_value(value: String) {
     #[test]
     fn reject_margin_string_case_expr_type_mismatch() {
         let mut module = compile_string_case_module();
-        let (type_, _, _) =
-            super::super::expect_case_statement_mut(&mut module.definitions.functions[0].body[0]);
+        let (type_, _, _) = super::super::super::expect_case_statement_mut(
+            &mut module.definitions.functions[0].body[0],
+        );
         *type_ = type_::bit_array();
         assert_eq!(plan_module(module), Err(case_branch_return_type_mismatch()));
